usb3: uint32_t transfer lengths and xHCI de-init on unplug

+++ src/class/msc/msc_host.c
@@ -319,7 +319,7 @@ bool msch_xfer_cb(uint8_t dev_addr, uint8_t ep_addr, xfer_result_t event, uint32
         // Data stage if any
         p_msc->stage = MSC_STAGE_DATA;
         uint8_t const ep_data = (cbw->dir & TUSB_DIR_IN_MASK) ? p_msc->ep_in : p_msc->ep_out;
-        TU_ASSERT(usbh_edpt_xfer(dev_addr, ep_data, p_msc->buffer, (uint16_t) cbw->total_bytes));
+        TU_ASSERT(usbh_edpt_xfer(dev_addr, ep_data, p_msc->buffer, (uint32_t) cbw->total_bytes));
         break;
       }
       TU_ATTR_FALLTHROUGH; // fallthrough to data stage
